1. page 的 start, limit 两个参数 改为 long类型。 2. ConditionSqlDialect 增加 randomQuery 随机查询方言方法。

+++ dbvisitor-core/src/main/java/net/hasor/dbvisitor/dal/execute/PreparedStatementExecute.java
@@ -57,8 +57,8 @@ public class PreparedStatementExecute extends AbstractStatementExecute<Object> {
 
         if (usingPage(executeInfo)) {
             PageSqlDialect dialect = executeInfo.pageDialect;
-            int position = executeInfo.pageInfo.getFirstRecordPosition();
-            int pageSize = executeInfo.pageInfo.getPageSize();
+            long position = executeInfo.pageInfo.getFirstRecordPosition();
+            long pageSize = executeInfo.pageInfo.getPageSize();
             boundSql = dialect.pageSql(sqlBuilder, position, pageSize);
             if (executeInfo.pageResult) {
                 countSql = dialect.countSql(sqlBuilder);
